backend/src/ticket/requirements.txt
openai; python_version>"3.6"
qdrant-client; python_version>"3.6"
python-dotenv